AccessNavigator_demo/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
databricks-sdk>=0.81.0
databricks-sql-connector>=3.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
openai>=1.52.0
requests>=2.31.0
